try to capture a metric for the start and end of the http request
