+++ packages/lesswrong/server/vulcan-lib/query.ts
@@ -8,7 +8,7 @@ import { localeSetting } from '../../lib/publicSettings';
 import { getExecutableSchema } from './apollo-server/initGraphQL';
 import { getCollectionsByName, generateDataLoaders } from './apollo-server/context';
 import { getAllRepos } from '../repos';
-import { generateTraceId } from '../perfMetrics';
+import { generateTraceId, openPerfMetric } from '../perfMetrics';
 
 function writeGraphQLErrorToStderr(errors: readonly GraphQLError[])
 {
@@ -56,7 +56,6 @@ export const createAnonymousContext = (options?: Partial<ResolverContext>): Reso
     ...getCollectionsByName(),
     ...generateDataLoaders(),
     ...options,
-    traceId: generateTraceId()
   };
   
   return queryContext;
